Change floorNumber to float in OsmLevel.

+++ application/src/main/java/org/opentripplanner/osm/issues/FloorNumberUnknownGuessedFromAltitude.java
@@ -5,7 +5,7 @@ import org.opentripplanner.osm.model.OsmEntity;
 
 public record FloorNumberUnknownGuessedFromAltitude(
   String layer,
-  Integer floorNumber,
+  Float floorNumber,
   OsmEntity entity
 )
   implements DataImportIssue {
